requirements.txt
asgiref==3.5.2
Django==4.1.2
django-braces==1.15.0
django-chartjs==2.3.0
django-cleanup==5.2.0
django-crispy-forms==1.13.0
djangorestframework==3.14.0
pytz==2021.3
sqlparse==0.4.2
